Add new customCleanEnv, with new tests for emulating non-strict mode
